requirements.txt
azure-functions
pyodbc
azure.storage.blob
pandas
xlrd


